feat(typing): Move `Optional` export to `typing`
Updates all references, docs, `ruff` rule
