+++ altair/vegalite/v5/schema/mixins.py
@@ -22,7 +22,7 @@ else:
 
 # ruff: noqa: F405
 if TYPE_CHECKING:
-    from altair.utils.schemapi import Optional
+    from altair.typing import Optional
 
     from ._typing import *  # noqa: F403
