+++ pyproject.toml
@@ -376,7 +376,12 @@ split-on-trailing-comma = false
 
 [tool.ruff.lint.flake8-tidy-imports.banned-api]
 # https://docs.astral.sh/ruff/settings/#lint_flake8-tidy-imports_banned-api
-"typing.Optional".msg = "Use `Union[T, None]` instead.\n`typing.Optional` is likely to be confused with `altair.Optional`, which have a similar but different semantic meaning.\nSee https://github.com/vega/altair/pull/3449"
+"typing.Optional".msg = """
+Use `Union[T, None]` instead.
+`typing.Optional` is likely to be confused with `altair.typing.Optional`, \
+which have a similar but different semantic meaning.
+See https://github.com/vega/altair/pull/3449
+"""
 
 [tool.ruff.lint.per-file-ignores]
 # Only enforce type annotation rules on public api
